Rename config file to match to service name
