+++ src/config/pixel.ts
